+++ test/unit/chromic_pdf/pdf/connection_test.exs
@@ -25,13 +25,7 @@ defmodule ChromicPDF.ConnectionTest do
 
   describe "initialization" do
     test "it spawns Chrome and initializes its state" do
-      opts = [discard_stderr: false, no_sandbox: true, chrome_args: "--foo"]
-      expect(ChromeMock, :spawn, fn ^opts -> {:ok, @port} end)
-      assert init({self(), opts}) == {:ok, new_state()}
-    end
-
-    test "it spawns custom Chrome executeable and initializes its state" do
-      opts = [discard_stderr: false, no_sandbox: true, chrome_executeable: "/cutom/chrome", chrome_args: "--foo"]
+      opts = [discard_stderr: false, no_sandbox: true, chrome_executable: "/custom/chrome", chrome_args: "--foo"]
       expect(ChromeMock, :spawn, fn ^opts -> {:ok, @port} end)
       assert init({self(), opts}) == {:ok, new_state()}
     end